requirements.txt
asgiref==3.9.1
certifi==2025.8.3
cloudinary==1.44.1
Django==5.2.5
django-ckeditor==6.7.3
django-js-asset==3.1.2
django-tinymce==4.1.0
django-unfold==0.65.0
django-widget-tweaks==1.5.0
mysqlclient==2.2.7
pillow==11.3.0
six==1.17.0
sqlparse==0.5.3
urllib3==2.5.0
